# artifacts/abyar/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OtpRequest {
  phone: string;
}

export interface OtpRequestResponse {
  accepted: boolean;
  devMode?: boolean;
}

export interface OtpVerifyRequest {
  phone: string;
  code: string;
}

export interface AuthUser {
  id: string;
  phone: string;
  /** @nullable */
  name?: string | null;
}

export interface AuthSession {
  token: string;
  user: AuthUser;
}

export interface WaterLog {
  id: string;
  amountMl: number;
  loggedAt: string;
}

export interface ReminderSettings {
  enabled: boolean;
  startTime: string;
  endTime: string;
  intervalMinutes: number;
  quietHoursEnabled: boolean;
  quietHoursStart: string;
  quietHoursEnd: string;
}

export type SyncStateProfile = {
  name: string;
  dailyGoalMl: number;
};

export interface SyncState {
  profile: SyncStateProfile;
  reminder: ReminderSettings;
  logs: WaterLog[];
}

export interface PartnerConnectRequest {
  inviteCode: string;
}

export interface PartnerSharingUpdate {
  shareProgress: boolean;
  shareLastDrink: boolean;
  shareHistory: boolean;
}

export type PartnerConnectionStatus = typeof PartnerConnectionStatus[keyof typeof PartnerConnectionStatus];


export const PartnerConnectionStatus = {
  pending: 'pending',
  active: 'active',
} as const;

export type PartnerConnectionRole = typeof PartnerConnectionRole[keyof typeof PartnerConnectionRole];


export const PartnerConnectionRole = {
  owner: 'owner',
  partner: 'partner',
} as const;

export interface PartnerConnection {
  id: string;
  status: PartnerConnectionStatus;
  role: PartnerConnectionRole;
  /** @nullable */
  partnerName?: string | null;
  /** @nullable */
  inviteCode?: string | null;
  canManageSharing: boolean;
  shareProgress: boolean;
  shareLastDrink: boolean;
  shareHistory: boolean;
}

export interface PartnerProgress {
  totalMl: number;
  goalMl: number;
  percent: number;
}

export interface PartnerSharedData {
  /** @nullable */
  partnerName: string | null;
  progress: PartnerProgress | null;
  lastDrink: WaterLog | null;
  history: WaterLog[];
}