# artifacts/abyar/lib/api-zod/src/generated/types/partnerConnection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerConnectionRole } from './partnerConnectionRole';
import type { PartnerConnectionStatus } from './partnerConnectionStatus';

export interface PartnerConnection {
  id: string;
  status: PartnerConnectionStatus;
  role: PartnerConnectionRole;
  /** @nullable */
  partnerName?: string | null;
  /** @nullable */
  inviteCode?: string | null;
  canManageSharing: boolean;
  shareProgress: boolean;
  shareLastDrink: boolean;
  shareHistory: boolean;
}