# artifacts/abyar/lib/api-zod/src/generated/types/partnerProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerProgress {
  totalMl: number;
  goalMl: number;
  percent: number;
}