# artifacts/abyar/lib/api-zod/src/generated/types/reminderSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReminderSettings {
  enabled: boolean;
  startTime: string;
  endTime: string;
  intervalMinutes: number;
  quietHoursEnabled: boolean;
  quietHoursStart: string;
  quietHoursEnd: string;
}