# artifacts/abyar/lib/api-zod/src/generated/types/waterLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaterLog {
  id: string;
  amountMl: number;
  loggedAt: string;
}